Move mapping logic to a new type `FeedItemsMapper`
